pyproject.toml
[tool.poetry]
name = "strawberry-subscription"
version = "0.1.0"
description = ""
authors = ["BurnySc2 <gamingburny@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {extras = ["debug-server"], version = "^0.86.1"}
fastapi = "^0.70.0"
loguru = "^0.5.3"
uvicorn = {extras = ["standard"], version = "^0.15.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
